+++ example/getDistance/getDistance.ino
@@ -18,7 +18,6 @@ uint16_t actDistance;
 void setup(void)
 {
   Wire.begin();
-
   Serial.begin(9600);
 
   while (sensor.begin() != true){
@@ -30,7 +29,7 @@ void setup(void)
 void loop(void)
 {
   sensor.startRanging();
-  int distance = sensor.getDistance();
+  uint16_t distance = sensor.getDistance();
   sensor.stopRanging();
 
   Serial.print("Distance(mm): ");
